+++ lib/src/controls/inputs/buttons/theme.dart
@@ -250,22 +250,16 @@ class ButtonThemeData with Diagnosticable {
   }
 
   static Color uncheckedInputColor(ThemeData style, Set<ButtonStates> states) {
-    // The opacity is 0 because, when transitioning between [Colors.transparent]
-    // and the actual color gives a weird effect
     if (style.brightness == Brightness.light) {
       if (states.isDisabled) return style.disabledColor;
-      if (states.isPressing) return Colors.grey[70];
-      if (states.isHovering) return Colors.grey[40];
-      return Colors.grey[40].withOpacity(0);
+      if (states.isPressing) return const Color(0xFF221D08).withOpacity(0.255);
+      if (states.isHovering) return const Color(0xFF221D08).withOpacity(0.075);
+      return Colors.transparent;
     } else {
       if (states.isDisabled) return style.disabledColor;
-      if (states.isPressing) return Colors.grey[130];
-      if (states.isHovering) return Colors.grey[150];
-      return Colors.grey[150].withOpacity(0);
+      if (states.isPressing) return const Color(0xFFFFF3E8).withOpacity(0.285);
+      if (states.isHovering) return const Color(0xFFFFF3E8).withOpacity(0.12);
+      return Colors.transparent;
     }
   }
-
-  // static Color iconButtonColor(ThemeData style) {
-
-  // }
 }